requirements.txt
# Core ML frameworks
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
xformers==0.0.22.post7

# Diffusion and transformers
diffusers==0.33.1
transformers==4.39.3
huggingface_hub==0.30.2
accelerate==0.24.0
peft==0.7.1

# Image processing
opencv-python==4.11.0.86
pillow==11.2.1
realesrgan==0.3.0
basicsr==1.4.2
controlnet-aux==0.0.9

# Utilities
numpy==1.26.4
omegaconf==2.3.0
setuptools==70.0.0
